fix(produto): Corrige dtos e repository

+++ src/main/java/com/gerenciamento/loja/repository/ProdutoRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProdutoRepository extends JpaRepository<ProdutoEntity,Integer> {
+public interface ProdutoRepository extends JpaRepository<ProdutoEntity,Long> {
 }
